Fix import circle deps issue
